refactor: delete MemberDao
