+++ src/main/java/gift/api/wishlist/repository/WishRepository.java
@@ -1,6 +1,5 @@
 package gift.api.wishlist.repository;
 
-import gift.api.member.domain.Member;
 import gift.api.wishlist.domain.Wish;
 import gift.api.wishlist.domain.WishId;
 import org.springframework.data.domain.Page;
@@ -10,5 +9,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface WishRepository extends JpaRepository<Wish, WishId> {
-    Page<Wish> findAllByMember(Member member, Pageable pageable);
+    Page<Wish> findAllByMemberId(Long member_id, Pageable pageable);
 }
